// cxh_pulse/frontend/src/types/component.types.ts
export type LocationOption = {
  value: string;
  label: string;
  type: 'level1' | 'level2' | 'level3';
  group: string;
  id: string;
  parentId?: string;
};